llm/pyproject.toml
[project]
name = "llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.5.0",
    "huggingface-hub[cli]>=0.30.1",
    "pillow>=11.1.0",
    "vllm==0.8.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm"]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
